The numerical issues are sorted -- updated method of calculating the quadrature points.

+++ reyna/DGFEM/two_dimensional/_auxilliaries/assembly/full_assembly.py
@@ -125,10 +125,13 @@ def int_localstiff(nodes: np.ndarray,
         c_inv_2 = min(k_2_area / abs_k_b_2, polydegree ** 2)
         sigma = sigma_D * lambda_dot * polydegree ** 2 * (2 * De) * max(c_inv_1 / k_1_area, c_inv_2 / k_2_area)
 
-        auxiliary_sigma_1 = np.zeros((dim_elem, dim_elem))
-        auxiliary_sigma_2 = np.zeros((dim_elem, dim_elem))
-        auxiliary_sigma_3 = np.zeros((dim_elem, dim_elem))
-        auxiliary_sigma_4 = np.zeros((dim_elem, dim_elem))
+        # Assuming not p-coverable
+        # sigma = sigma_D * lambda_dot * polydegree ** 2 * (2 * De) * max(1.0 / abs_k_b_1, 1.0 / abs_k_b_2)
+
+        auxiliary_sigma_1 = np.zeros((dim_elem, dim_elem), dtype=np.float64)
+        auxiliary_sigma_2 = np.zeros((dim_elem, dim_elem), dtype=np.float64)
+        auxiliary_sigma_3 = np.zeros((dim_elem, dim_elem), dtype=np.float64)
+        auxiliary_sigma_4 = np.zeros((dim_elem, dim_elem), dtype=np.float64)
 
         gradtensor_leg_array = np.stack(
             (tensor_gradtensor_leg(P_Qpoints, m1, h1, Lege_ind),
